test(web): split smoke suite by workflow
