+++ tests/web-smoke/varlens/auth.api.cy.ts
@@ -1,0 +1,11 @@
+import { expectCoreCaseApisAvailable, expectCurrentUserApiAvailable } from '../support/api'
+import { beforeEachAuthenticatedSmoke } from '../support/auth'
+
+describe('VarLens authenticated API smoke', () => {
+  beforeEachAuthenticatedSmoke()
+
+  it('returns the current user and core case APIs after login', () => {
+    expectCurrentUserApiAvailable()
+    expectCoreCaseApisAvailable()
+  })
+})
